Add `character_set` to dump.log
